feat: add description and update VIP number for VIP-369
